backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
cerebras-cloud-sdk>=1.50.0
aiofiles==23.2.1
requests==2.31.0
pillow==10.1.0
python-dateutil==2.8.2
fastapi-cors==0.0.6
